Split d3chart/utils.js
