--- packages/components/src/chart/d3chart/test/fixtures/dummy-hour.js
@@ -1,25 +1,0 @@
-/** @format */
-
-// /**
-//  * /* eslint-disable quote-props
-//  *
-//  * @format
-//  */
-
-export default [
-	{
-		date: '2018-08-01T00:00:00',
-		'Custom (Aug 1, 2018)': { label: '2018-08-01 00:00', value: 58929.99 },
-		'Previous Period (Jul 31, 2018)': { label: '2018-07-31 00:00', value: 160130.74000000002 },
-	},
-	{
-		date: '2018-08-01T01:00:00',
-		'Custom (Aug 1, 2018)': { label: '2018-08-01 01:00', value: 3805.56 },
-		'Previous Period (Jul 31, 2018)': { label: '2018-07-31 01:00', value: 0 },
-	},
-	{
-		date: '2018-08-01T02:00:00',
-		'Custom (Aug 1, 2018)': { label: '2018-08-01 02:00', value: 3805.56 },
-		'Previous Period (Jul 31, 2018)': { label: '2018-07-31 02:00', value: 0 },
-	},
-];
