+++ packages/components/src/chart/d3chart/utils/breakpoints.js
@@ -1,0 +1,2 @@
+export const smallBreak = 783;
+export const wideBreak = 1365;
