+++ packages/components/src/chart/d3chart/utils/test/fixtures/dummy-ordered-dates.js
@@ -1,0 +1,9 @@
+/** @format */
+export default [
+	'2018-05-30T00:00:00',
+	'2018-05-31T00:00:00',
+	'2018-06-01T00:00:00',
+	'2018-06-02T00:00:00',
+	'2018-06-03T00:00:00',
+	'2018-06-04T00:00:00',
+];
