+++ packages/components/src/chart/d3chart/utils/test/fixtures/dummy-orders.js
@@ -1,11 +1,4 @@
 /** @format */
-
-/**
- * /* eslint-disable quote-props
- *
- * @format
- */
-
 export default [
 	{
 		date: '2018-05-30T00:00:00',
